Refactor Item and ItemList to add items using actions/reducers
