+++ client/src/sections/Fridge/ItemList/ItemList.tsx
@@ -1,4 +1,4 @@
-import React, {useState, useCallback} from 'react';
+import React from 'react';
 import {useSelector} from 'react-redux';
 import {Card} from '@shopify/polaris';
 import {selectItems} from 'store/item/selectors';
@@ -7,27 +7,7 @@ import {Item, AddItemModal} from './components';
 import './ItemList.css';
 
 export default function ItemList() {
-  const [items, setItems] = useState(useSelector(selectItems));
-  const [value, setValue] = useState('');
-
-  const handleTextChange = useCallback((newValue) => {
-    setValue(newValue);
-  }, []);
-
-  const handleAddItem = useCallback(() => {
-    const newItem = {
-      // this will come from the database once setup
-      id: items.length,
-      name: value,
-      // this will come from the database once setup
-      category: 'Fruits',
-      count: 1,
-      expiry: new Date().toLocaleDateString(),
-    };
-
-    setItems([...items, newItem]);
-    setValue('');
-  }, [items, value]);
+  const items = useSelector(selectItems);
 
   const itemListMarkup = items.map(({name, expiry}, id) => {
     return (
@@ -41,11 +21,7 @@ export default function ItemList() {
     <div className='item-list'>
       <div className='item'>
         <Card sectioned>
-          <AddItemModal
-            value={value}
-            onValueChange={handleTextChange}
-            onAdd={handleAddItem}
-          />
+          <AddItemModal />
         </Card>
       </div>
       {itemListMarkup}
